Move several definitions to `runtime/job.{hpp,src}`
